Moved to template recursion for sin/cos
This commit transitions hx::sin and hx::cos into more
traditional recursive struct templates in an attempt
to reduce the number of constexpr steps taken when
building hx::fft.

As suspected, the major constexpr culprit is actually
hx::fft_shuffle. More work is needed to find a proper
solution for this one.