Added hx::vector, for vector views
The hx::vector class template is a razor-thin wrapper around
the raw data pointer of an hx::array that implements strided
subscript and shift operators.

An example of applying an hx::fft::forward to each column
of a matrix is given in test.cc. However, this example code
raises a point about the next deficiency in the codebase...

hx::index has elegant {pre,post}-{inc,dec}rement operators,
each of which has an intuitive meaning for how the index
is changed. (i.e. idx++ increments the last index first,
while ++idx increments the first index first, etc...)

It seems that a battery of operator+=() and operator-=()
are needed, but which take either std::size_t or a special
'fat-free' type (e.g. "hx::dim<std::size_t>"?). The size_t
+= works as if the user called the associated ++ n times,
while the hx::dim += skips that dimension?

More thought required.