Towards array-valued expression support
This commit begins efforts to encode array-valued
expressions as types. For example, we want the
[element-wise] expression:

 z = 2 * x + hx::scalar<1>{ 1, 2 } * y - 3.14;

to defer *all* computations until the operator=()
is reached, thus avoiding copies of intermediate
arrays. This functionality is similar to that of
Eigen, but we're doing it for arbitrarily-sized
arrays and we're only concerned with element-wise
operations (i.e. not linear-algebraic ones).