Towards TMP-based processing functionality
This commit introduces a very limited set of functions
for processing arrays using a high-level declarative
syntax. Specifically, method chaining:

  using X = hx::array<hx::scalar<1>, 512>;
  auto x = std::make_shared<X>();

  auto f = hx::proc::node(x).zerofill().fft().real();
  auto y = f();

The hx::proc::node class exposes a set of member functions,
each of which returns a new node with an associated processor
type that performs the requested functionality.

Much more work is needed before this is fully ironed out,
even more before it's of any practical utility, but it
seems to be moving in the right direction.